introduction of url-prefixes and defaults
