+++ src/cryptoadvance/specter/server_endpoints/controller.py
@@ -25,6 +25,7 @@ from .filters import filters_bp
 app.register_blueprint(filters_bp)
 
 # Setup specter endpoints
+from .welcome import welcome_endpoint
 from .auth import auth_endpoint
 from .devices import devices_endpoint
 from .nodes import nodes_endpoint
@@ -38,15 +39,17 @@ from ..rpc import RpcError
 # Services live in their own separate path
 from cryptoadvance.specter.services.controller import services_endpoint
 
-app.register_blueprint(auth_endpoint, url_prefix="/auth")
-app.register_blueprint(devices_endpoint, url_prefix="/devices")
-app.register_blueprint(nodes_endpoint, url_prefix="/nodes")
-app.register_blueprint(price_endpoint, url_prefix="/price")
-app.register_blueprint(services_endpoint, url_prefix="/services")
-app.register_blueprint(settings_endpoint, url_prefix="/settings")
-app.register_blueprint(setup_endpoint, url_prefix="/setup")
-app.register_blueprint(wallets_endpoint, url_prefix="/wallets")
-app.register_blueprint(wallets_endpoint_api, url_prefix="/wallets")
+spc_prefix = app.config["SPECTER_URL_PREFIX"]
+app.register_blueprint(welcome_endpoint, url_prefix=f"{spc_prefix}/welcome")
+app.register_blueprint(auth_endpoint, url_prefix=f"{spc_prefix}/auth")
+app.register_blueprint(devices_endpoint, url_prefix=f"{spc_prefix}/devices")
+app.register_blueprint(nodes_endpoint, url_prefix=f"{spc_prefix}/nodes")
+app.register_blueprint(price_endpoint, url_prefix=f"{spc_prefix}/price")
+app.register_blueprint(services_endpoint, url_prefix=f"{spc_prefix}/services")
+app.register_blueprint(settings_endpoint, url_prefix=f"{spc_prefix}/settings")
+app.register_blueprint(setup_endpoint, url_prefix=f"{spc_prefix}/setup")
+app.register_blueprint(wallets_endpoint, url_prefix=f"{spc_prefix}/wallets")
+app.register_blueprint(wallets_endpoint_api, url_prefix=f"{spc_prefix}/wallets")
 
 rand = random.randint(0, 1e32)  # to force style refresh
 
@@ -65,7 +68,7 @@ def server_rpc_error(rpce):
         app.specter.wallet_manager.update()
     except SpecterError as se:
         flash(str(se), "error")
-    return redirect(url_for("about"))
+    return redirect(url_for("welcome_endpoint.about"))
 
 
 @app.errorhandler(SpecterError)
@@ -80,7 +83,7 @@ def server_specter_error(se):
         return redirect(request.url)
     # potentially avoiding http loops. Might be improvable but how?
     else:
-        return redirect(url_for("about"))
+        return redirect(url_for("welcome_endpoint.about"))
 
 
 @app.errorhandler(NotFound)
@@ -161,14 +164,18 @@ def selfcheck():
 
 
 @app.before_request
-def slow_request_detection():
+def slow_request_detection_start():
     """ """
     g.start = time()
 
 
 @app.after_request
-def after_request(response):
-    diff = time() - g.start
+def slow_request_detection_stop(response):
+    try:
+        diff = time() - g.start
+    except Exception as e:
+        app.logger.error(e)
+        return response
     if (
         (response.response)
         and (200 <= response.status_code < 300)
@@ -196,170 +203,14 @@ def inject_debug():
 
 ################ Specter global routes ####################
 @app.route("/")
-@login_required
 def index():
-    if request.args.get("mode"):
-        if request.args.get("mode") == "remote":
-            pass
-    notify_upgrade(app, flash)
-    if len(app.specter.wallet_manager.wallets) > 0:
-        if len(app.specter.wallet_manager.wallets) > 1:
-            return redirect(url_for("wallets_endpoint.wallets_overview"))
-        return redirect(
-            url_for(
-                "wallets_endpoint.wallet",
-                wallet_alias=app.specter.wallet_manager.wallets[
-                    app.specter.wallet_manager.wallets_names[0]
-                ].alias,
-            )
-        )
-
-    return redirect("about")
+    app.logger.info("Muuuh")
+    return redirect(url_for("welcome_endpoint.index"))
 
 
-@app.route("/about", methods=["GET", "POST"])
-@login_required
-def about():
-    notify_upgrade(app, flash)
-    if request.method == "POST":
-        action = request.form["action"]
-        if action == "cancelsetup":
-            app.specter.setup_status["stage"] = "start"
-            app.specter.reset_setup("bitcoind")
-            app.specter.reset_setup("torbrowser")
+if app.config["SPECTER_URL_PREFIX"] != "":
 
-    return render_template(
-        "base.jinja",
-        specter=app.specter,
-        rand=rand,
-        supported_languages=app.supported_languages,
-    )
-
-
-# TODO: Move all these below to REST API
-
-################ Utils ####################
-
-
-@app.route("/wallets_loading/", methods=["GET", "POST"])
-@login_required
-def wallets_loading():
-    return {
-        "is_loading": app.specter.wallet_manager.is_loading,
-        "loaded_wallets": [
-            app.specter.wallet_manager.wallets[wallet].alias
-            for wallet in app.specter.wallet_manager.wallets
-        ],
-        "failed_load_wallets": [
-            wallet["alias"] for wallet in app.specter.wallet_manager.failed_load_wallets
-        ],
-    }
-
-
-@app.route("/generatemnemonic/", methods=["GET", "POST"])
-@login_required
-def generatemnemonic():
-    return {
-        "mnemonic": generate_mnemonic(
-            strength=int(request.form["strength"]),
-            language_code=app.get_language_code(),
-        )
-    }
-
-
-################ RPC data utils ####################
-@app.route("/get_fee/<blocks>")
-@login_required
-def fees(blocks):
-    return app.specter.estimatesmartfee(int(blocks))
-
-
-@app.route("/get_txout_set_info")
-@login_required
-@app.csrf.exempt
-def txout_set_info():
-    res = app.specter.rpc.gettxoutsetinfo()
-    return res
-
-
-@app.route("/get_scantxoutset_status")
-@login_required
-@app.csrf.exempt
-def get_scantxoutset_status():
-    status = app.specter.rpc.scantxoutset("status", [])
-    app.specter.info["utxorescan"] = status.get("progress", None) if status else None
-    if app.specter.info["utxorescan"] is None:
-        app.specter.utxorescanwallet = None
-    return {
-        "active": app.specter.info["utxorescan"] is not None,
-        "progress": app.specter.info["utxorescan"],
-    }
-
-
-@app.route("/toggle_hide_sensitive_info/", methods=["POST"])
-@login_required
-@app.csrf.exempt  # might get called by a timeout in the browser --> csrf-issues
-def toggle_hide_sensitive_info():
-    try:
-        app.specter.update_hide_sensitive_info(
-            not app.specter.hide_sensitive_info, current_user
-        )
-        return {"success": True}
-    except Exception as e:
-        app.logger.warning(
-            "Failed to update sensitive info display settings. Exception: {}".format(e)
-        )
-    return {"success": False}
-
-
-@app.route("/bitcoin.pdf")
-@login_required
-def get_whitepaper():
-    if app.specter.chain == "main":
-        if not app.specter.info["pruned"]:
-            raw_tx = app.specter.rpc.getrawtransaction(
-                "54e48e5f5c656b26c3bca14a8c95aa583d07ebe84dde3b7dd4a78f4e4186e713",
-                False,
-                "00000000000000ecbbff6bafb7efa2f7df05b227d5c73dca8f2635af32a2e949",
-            )
-            outputs = raw_tx.split("0100000000000000")
-            pdf = ""
-            for output in outputs[1:-2]:
-                cur = 6
-                pdf += output[cur : cur + 130]
-                cur += 132
-                pdf += output[cur : cur + 130]
-                cur += 132
-                pdf += output[cur : cur + 130]
-            pdf += outputs[-2][6:-4]
-        else:
-            outputs_prun = app.specter.rpc.multi(
-                [
-                    (
-                        "gettxout",
-                        "54e48e5f5c656b26c3bca14a8c95aa583d07ebe84dde3b7dd4a78f4e4186e713",
-                        i,
-                    )
-                    for i in range(0, 946)
-                ]
-            )
-            pdf = ""
-            for output in outputs_prun[:-1]:
-                cur = 4
-                pdf += output["result"]["scriptPubKey"]["hex"][cur : cur + 130]
-                cur += 132
-                pdf += output["result"]["scriptPubKey"]["hex"][cur : cur + 130]
-                cur += 132
-                pdf += output["result"]["scriptPubKey"]["hex"][cur : cur + 130]
-            pdf += outputs_prun[-1]["result"]["scriptPubKey"]["hex"][4:-4]
-        res = make_response(unhexlify(pdf[16:-16]))
-        res.headers.set("Content-Disposition", "attachment")
-        res.headers.set("Content-Type", "application/pdf")
-        return res
-    else:
-        return render_template(
-            "500.jinja",
-            error=_(
-                "You need a mainnet node to retrieve the whitepaper. Check your node configurations."
-            ),
-        )
+    @app.route(f"{app.config['SPECTER_URL_PREFIX']}/")
+    def index_prefix():
+        app.logger.info("Muuuh")
+        return redirect(url_for("welcome_endpoint.index"))
