+++ src/cryptoadvance/specter/templates/500.jinja
@@ -18,7 +18,7 @@
             <h1>{{ _("Server error!") }}</h1>
             {{ _("Exception") }}: <b>{{ error }}</b><br><br>
             <div><br>
-                <a class="btn" href="{{ url_for('index') }}">{{ _("Back to the main page") }}</a>
+                <a class="btn" href="{{ url_for('welcome_endpoint.index') }}">{{ _("Back to the main page") }}</a>
             <br></div>
             {{ _("Details") }}:
             <pre><code>{{ traceback }}</code></pre>
