+++ src/cryptoadvance/specter/templates/500_timeout.jinja
@@ -18,7 +18,7 @@
             <h1>{{ _("Timeout Error!") }}</h1>
             {{ _("Exception") }}: <b>{{ error }}</b><br><br>
             <div><br>
-                <a class="btn" href="{{ url_for('index') }}">{{ _("Back to the main page") }}</a>
+                <a class="btn" href="{{ url_for('welcome_endpoint.index') }}">{{ _("Back to the main page") }}</a>
             <br></div>
             {{ _("Please have a look in the below log. There are effectively two ways to resolve this:") }}"
             <ul>
